Split liquidity supportability tests
